refactor: move w_ring_buffer to a standalone header
 - move w_ring_buffer from wolf::media::internal to wolf::system to be
   accessible as a general utility throughout the project.
 - rename w_ring_buffer to w_spsc_ring_buffer for clarity as it's now a
   generally available utility instead of just openal.

issue #420
